update security 2fa and forceLogout
